+++ src/rndartt/apalette.cpp
@@ -8,7 +8,7 @@ namespace {
 
 // Larger than the largest sum of three squared byte differences, 0x2fa43, so the first entry of
 // any non empty range wins the comparison.
-constexpr long kNoMatchDistance = 0x30000;
+constexpr long long kNoMatchDistance = 0x30000;
 
 constexpr unsigned int kChannelMask = 0xff;
 constexpr int kGreenShift = 8;
@@ -74,17 +74,17 @@ void APalette::SetEntriesRGB(const unsigned char *pRGB, int nFirst, int nCount) 
 
 // 0x00613f10
 int APalette::FindNearestEntry(unsigned int nColor, int nFirst, int nLast) const {
-    const long nRed = nColor & kChannelMask;
-    const long nGreen = (nColor >> kGreenShift) & kChannelMask;
-    const long nBlue = (nColor >> kBlueShift) & kChannelMask;
-    long nBestDistance = kNoMatchDistance;
+    const long long nRed = nColor & kChannelMask;
+    const long long nGreen = (nColor >> kGreenShift) & kChannelMask;
+    const long long nBlue = (nColor >> kBlueShift) & kChannelMask;
+    long long nBestDistance = kNoMatchDistance;
     int nBestIndex = 0;
     for (int i = nFirst; i <= nLast; ++i) {
         const unsigned int nEntry = mEntries[i];
-        const long nDeltaRed = nRed - (nEntry & kChannelMask);
-        const long nDeltaGreen = nGreen - ((nEntry >> kGreenShift) & kChannelMask);
-        const long nDeltaBlue = nBlue - ((nEntry >> kBlueShift) & kChannelMask);
-        const long nDistance =
+        const long long nDeltaRed = nRed - (nEntry & kChannelMask);
+        const long long nDeltaGreen = nGreen - ((nEntry >> kGreenShift) & kChannelMask);
+        const long long nDeltaBlue = nBlue - ((nEntry >> kBlueShift) & kChannelMask);
+        const long long nDistance =
             nDeltaRed * nDeltaRed + nDeltaGreen * nDeltaGreen + nDeltaBlue * nDeltaBlue;
         if (nDistance < nBestDistance) {
             nBestDistance = nDistance;
